Extract brokertrigger test package
This lets us define different topologies and is a little more explicit
in the test what's being setup
