+++ test/e2e/config/brokertrigger/trigger.yaml
@@ -1,0 +1,15 @@
+{{- range $index, $kref := .triggers -}}
+---
+apiVersion: eventing.knative.dev/v1
+kind: Trigger
+metadata:
+  name: {{ printf "%s-%d" $kref.Name $index }}
+  namespace: {{ $.namespace }}
+spec:
+  broker: testbroker
+  subscriber:
+    ref:
+      apiVersion: v1
+      kind: {{ $kref.Kind }}
+      name: {{ $kref.Name }}
+{{- end -}}
