--- test/e2e/config/direct/trigger.yaml
@@ -1,26 +1,0 @@
-# Copyright 2020 The Knative Authors
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     https://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-apiVersion: eventing.knative.dev/v1
-kind: Trigger
-metadata:
-  name: testtrigger
-  namespace: {{ .namespace }}
-spec:
-  broker: testbroker
-  subscriber:
-    ref:
-      apiVersion: v1
-      kind: Service
-      name: recorder
